refactor: Extract Checkbox into its own component
Since we have to use a workaround in order to have a custom checkbox we
are unable to use a class. Therefore we create a Checkbox component

Currently the Checkbox works for both true/false values as well as
custom ones, but the user will be unable to focus on the HTML element.

TODO: Make the Checkbox HTML accessible